requirements.txt
anthropic>=0.40.0
beautifulsoup4>=4.12.0
click>=8.0.0
google-auth>=2.0.0
google-api-python-client>=2.0.0
httpx>=0.27.0
lxml>=5.0.0
ollama>=0.3.0
python-dotenv>=1.0.0
requests>=2.28.0
rich>=13.0.0
aiofiles>=23.0.0
textstat>=0.7.0
